requirements.txt
albumentations==2.0.6
comet_ml==3.49.9
coremltools==8.3.0
Flask==3.1.0
GitPython==3.1.44
ipython==8.12.3
matplotlib==3.10.1
mss==10.0.0
nncf==2.16.0
numpy==2.2.5
onnx==1.17.0
onnxruntime==1.21.1
onnxslim==0.1.51
opencv_python==4.11.0.86
openvino==2025.1.0
paddle==1.0.2
pafy==0.5.5
pandas==2.2.3
Pillow==11.2.1
psutil==7.0.0
pycocotools==2.0.8
PyYAML==6.0.2
Requests==2.32.3
scipy==1.15.3
seaborn==0.13.2
tensorflow==2.19.0
tensorflowjs==4.22.0
tensorrt==10.10.0.31
tflite_runtime==2.14.0
tflite_support==0.4.4
thop==0.1.1.post2209072238
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
tritonclient==2.56.0
ultralytics==8.3.129
x2paddle==1.6.0
